fixes according review
