+++ src/components/Orders/Orders.module.css
@@ -1,30 +1,12 @@
 .orders {
-    margin: 120px auto 0;
+    margin: 40px auto 0;
     width: 1280px;
     display: flex;
 }
 
 .menu {
     width: 320px;
-}
-
-.links {
-    padding: 0 0 0 20px;
-}
-
-.link {
-    text-decoration: none;
-    height: 64px;
-    list-style-type: none;
-}
-
-.navlink {
-    color: #8585ad;
-    text-decoration: none;
-}
-
-.navlink_active {
-    color: white;
+    margin-top:80px;
 }
 
 .description {
