+++ src/components/Profile/Profile.module.css
@@ -8,31 +8,8 @@
     width: 320px;
 }
 
-.links {
-    padding: 0 0 0 20px;
-}
-
-.link {
-    text-decoration: none;
-    height: 64px;
-    list-style-type: none;
-}
-
-.navlink {
-    color: #8585ad;
-    text-decoration: none;
-}
-
-.navlink_active {
-    color: white;
-}
-
-.logout {
-    background-color: transparent;
-    outline: none;
-    border: none;
-    cursor: pointer;
-    color: #8585ad;
+.description {
+    width: 320px;
 }
 
 .form {
